+++ src/routes/index.js
@@ -2,7 +2,7 @@ import Dashboard from 'layouts/Dashboard/Dashboard.js';
 import LoginPage from 'layouts/Login/LoginPage';
 import ConfirmUser from 'layouts/ConfirmUser/ConfirmUser'
 import ResetPassword from 'layouts/ResetPassowrd/ResetPassword';
-import './fullscreen'
+// import './fullscreen'
 
 const indexRoutes = [
 	{ path: '/login', component: LoginPage },
